fix: :bug: Fixed all change requests
